Add files via upload
React app project/,MERN based, where users can find cocktail bars, breweries or wineries to visit. Users are directed to a list of bars to choose from and upon clicking on a bar, they are able to find details pertaining to the bar with features such as a picture of the bar, it's address, whether the bar takes reservations, allows children and the ambience of the bar. (future features will include users to read and add reviews for each bar). Incorrect info or change in business hours/address? No problem, a user can update the bar info to reflect the most current info. Bar is out of business? Users can also delete a bar if it is no longer in business. A new spot just open? Users can also add a new bar.